env.example modified
hey abhijeet, we dont have to mention our details in .env as well as in .yml file.
we have to mention our details in only in .env file and also we dont have to commit .env file.

one more thing in .env.example file we have to mention DATASOURCE_URL also as we will fetch that dynamically.